setup_startup_scripts ported to toonboom.lib, exposed in __init__.py. Also fixed typo for thread.daemon
